@@ -1,389 +1,419 @@
+using System;
+using System.Buffers;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
+using System.Linq;
+using System.Net.Http;
+using System.Net.Http.Json;
+using System.Net.WebSockets;
+using System.Text.Json.Nodes;
+using System.Threading;
+using System.Threading.Tasks;
+using Bloom.Events;
+using Bloom.Parsing;
+using Bloom.Payloads;
+using Bloom.Playback;
+using Bloom.Searching;
+using Discord;
+using Discord.WebSocket;
+
 namespace Bloom;
 
+/// <summary>
+/// Represents a node that connects to the Lavalink server.
+/// </summary>
 public sealed class BloomNode : IAsyncDisposable
 {
-    private const string LoadTracksEndpoint = "/v3/loadtracks?identifier={0}";
+    private const string LoadTracksEndpoint = "loadtracks?identifier={0}";
+    private const string UpdatePlayerEndpoint = "sessions/{0}/players/{1}";
 
     private readonly DiscordSocketClient _discordClient;
-    private readonly BloomConfig _config;
-    private int _reconnectAttemps;
-    private int _reconnectDelay;
-    private readonly ClientWebSocket _webSocket;
-    private readonly HttpClient _httpClient;
-    private readonly CancellationTokenSource _connectionTokenSource;
+    private readonly BloomConfiguration _configuration;
+    private readonly ClientWebSocket _webSocketClient;
+    private readonly HttpClient _restClient;
     private readonly Dictionary<ulong, BloomPlayer> _players;
+
+    private string _sessionId;
+    private Task? _receivingTask;
     private CancellationTokenSource? _cancellationTokenSource;
 
+    /// <summary>
+    /// Occurs when an exception is thrown by the node.
+    /// </summary>
     public event NodeExceptionEvent? NodeException;
+
+    /// <summary>
+    /// Occurs when a track starts playing.
+    /// </summary>
     public event TrackStartEvent? TrackStarted;
+
+    /// <summary>
+    /// Occurs when a track ends playing.
+    /// </summary>
     public event TrackEndEvent? TrackEnded;
+
+    /// <summary>
+    /// Occurs when an exception is thrown by a track.
+    /// </summary>
     public event TrackExceptionEvent? TrackException;
+
+    /// <summary>
+    /// Occurs when a track gets stuck.
+    /// </summary>
     public event TrackStuckEvent? TrackStucked;
 
-    public bool IsConnected { get; private set; }
+    /// <summary>
+    /// Gets a value indicating whether the node is connected to the remote server.
+    /// </summary>
+    public bool IsConnected => _webSocketClient.State is WebSocketState.Open;
 
-    internal string SessionId { get; private set; }
-
-    public BloomNode(DiscordSocketClient discordClient, BloomConfig config)
+    /// <summary>
+    /// Initializes a new instance of the <see cref="BloomNode"/> class.
+    /// </summary>
+    /// <param name="discordClient">The Discord client.</param>
+    /// <param name="configuration">The configuration of the node.</param>
+    public BloomNode(DiscordSocketClient discordClient, BloomConfiguration configuration)
     {
         _discordClient = discordClient;
-        _config = config;
-        _webSocket = new ClientWebSocket();
-        _httpClient = new HttpClient();
-        _connectionTokenSource = new CancellationTokenSource();
-        _players = new Dictionary<ulong, BloomPlayer>();
-        IsConnected = false;
-        SessionId = string.Empty;
+        _configuration = configuration;
+        _webSocketClient = new ClientWebSocket();
+        _restClient = new HttpClient();
+        _players = [];
+        _sessionId = string.Empty;
     }
 
+    /// <inheritdoc/>
     public async ValueTask DisposeAsync()
     {
         if (IsConnected)
             await DisconnectAsync();
 
-        _connectionTokenSource?.Dispose();
-        _webSocket.Dispose();
+        _webSocketClient.Dispose();
+        _restClient.Dispose();
     }
 
-    public async Task ConnectAsync()
+    /// <summary>
+    /// Returns the player of the specified guild.
+    /// </summary>
+    /// <param name="guild">The guild where the player is belong to.</param>
+    /// <returns>The player that is found in the specified guild.</returns>
+    public bool TryGetPlayer(IGuild guild, [MaybeNullWhen(false)] out BloomPlayer player)
     {
-        if (_webSocket.State is WebSocketState.Open)
-            throw new InvalidOperationException("The node already in open state");
+        return _players.TryGetValue(guild.Id, out player);
+    }
+
+    /// <summary>
+    /// Connects the node to the remote server.
+    /// </summary>
+    /// <param name="cancellationToken">The optional cancellation token.</param>
+    /// <returns>A <see cref="ValueTask"/> representing the asynchronous operation.</returns>
+    /// <exception cref="InvalidOperationException"></exception>
+    public async ValueTask ConnectAsync(CancellationToken cancellationToken = default)
+    {
+        ThrowIfConnected();
 
         _discordClient.VoiceServerUpdated += VoiceServerUpdated;
         _discordClient.UserVoiceStateUpdated += UserVoiceStateUpdated;
 
-        _webSocket.Options.SetRequestHeader("User-Id", _discordClient.CurrentUser.Id.ToString());
-        _webSocket.Options.SetRequestHeader("Client-Name", $"{nameof(Bloom)}/{typeof(BloomNode).Assembly.GetName().Version}");
-        _webSocket.Options.SetRequestHeader("Num-Shards", _config.ShardCount.ToString());
-        _webSocket.Options.SetRequestHeader("Authorization", _config.Authorization);
+        _webSocketClient.Options.SetRequestHeader("User-Id", _discordClient.CurrentUser.Id.ToString());
+        _webSocketClient.Options.SetRequestHeader("Client-Name", $"{nameof(Bloom)}/{typeof(BloomNode).Assembly.GetName().Version}");
+        _webSocketClient.Options.SetRequestHeader("Num-Shards", _configuration.ShardCount.ToString());
+        _webSocketClient.Options.SetRequestHeader("Authorization", _configuration.Authorization);
 
-        _httpClient.DefaultRequestHeaders.Add("Authorization", _config.Authorization);
+        _restClient.BaseAddress = new Uri(_configuration.RestEndpoint);
+        _restClient.DefaultRequestHeaders.Add("Authorization", _configuration.Authorization);
 
-        async Task VerifyConnectionAsync(Task task)
+        async ValueTask EstablishConnectionAsync(int reconnectAttemps = 1)
         {
-            if (task.Exception is not null)
+            if (reconnectAttemps > _configuration.ReconnectAttemps)
+                throw new InvalidOperationException("The node could not establish a connection");
+
+            try
             {
-                await ReconnectAsync();
-                return;
+                await _webSocketClient.ConnectAsync(new Uri(_configuration.WebSocketEndpoint), cancellationToken);
+            }
+            catch (OperationCanceledException)
+            {
+            }
+            catch (WebSocketException ex)
+            {
+                NodeException?.Invoke(new NodeExceptionEventArgs(ex.Message));
+                await Task.Delay(_configuration.ReconnectDelayInMiliseconds * reconnectAttemps, cancellationToken);
+                await EstablishConnectionAsync(reconnectAttemps + 1);
+            }
+        }
+
+        await EstablishConnectionAsync();
+
+        _cancellationTokenSource = new CancellationTokenSource();
+        _receivingTask = Task.Run(StartReceivingAsync, cancellationToken);
+    }
+
+    /// <summary>
+    /// Disconnects the node from the remote server.
+    /// </summary>
+    /// <param name="cancellationToken">The optional cancellation token.</param>
+    /// <returns>A <see cref="ValueTask"/> representing the asynchronous operation.</returns>
+    /// <exception cref="InvalidOperationException"></exception>
+    public async ValueTask DisconnectAsync(CancellationToken cancellationToken = default)
+    {
+        ThrowIfNotConnected();
+
+        try
+        {
+            _cancellationTokenSource?.Cancel();
+            _cancellationTokenSource?.Dispose();
+            _cancellationTokenSource = null;
+
+            if (_receivingTask is not null)
+            {
+                await _receivingTask;
+                _receivingTask = null;
             }
 
-            IsConnected = true;
-            await StartReceivingAsync();
-        }
+            await _webSocketClient.CloseAsync(WebSocketCloseStatus.NormalClosure, "Disconnect Request", cancellationToken);
 
-        try
-        {
-            await _webSocket
-                .ConnectAsync(new Uri(_config.Endpoint), _connectionTokenSource.Token)
-                .ContinueWith(VerifyConnectionAsync);
+            _discordClient.VoiceServerUpdated -= VoiceServerUpdated;
+            _discordClient.UserVoiceStateUpdated -= UserVoiceStateUpdated;
         }
-        catch (Exception exception)
+        catch (OperationCanceledException)
         {
-            if (exception is not ObjectDisposedException)
-                return;
-
-            NodeException?.Invoke(new NodeExceptionEventArgs(exception.Message));
-            await ReconnectAsync();
+        }
+        catch (Exception ex)
+        {
+            NodeException?.Invoke(new NodeExceptionEventArgs(ex.Message));
         }
     }
 
-    public async Task DisconnectAsync()
+    /// <summary>
+    /// Joins a new player to the specified voice channel.
+    /// </summary>
+    /// <param name="voiceChannel">The voice channel where the player will be connected to.</param>
+    /// <param name="textChannel">The text channel where the player will be connected to.</param>
+    /// <returns>The player that is connected to the voice channel.</returns>
+    /// <exception cref="InvalidOperationException"></exception>
+    public async ValueTask<BloomPlayer> JoinAsync(IVoiceChannel voiceChannel, IMessageChannel textChannel)
     {
-        if (_webSocket.State is not WebSocketState.Open)
-            throw new InvalidOperationException("The node is not in open state");
-
-        try
-        {
-            await _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, "Disconnect Request", _connectionTokenSource.Token);
-        }
-        catch (Exception exception)
-        {
-            NodeException?.Invoke(new NodeExceptionEventArgs(exception.Message));
-        }
-        finally
-        {
-            IsConnected = false;
-            _connectionTokenSource.Cancel(false);
-        }
-    }
-
-    public BloomPlayer? GetPlayer(IGuild guild)
-    {
-        _ = _players.TryGetValue(guild.Id, out BloomPlayer? player);
-        return player;
-    }
-
-    public bool HasPlayer(IGuild guild)
-    {
-        return _players.ContainsKey(guild.Id);
-    }
-
-    public async Task<BloomPlayer> JoinAsync(IVoiceChannel voiceChannel, IMessageChannel textChannel)
-    {
-        if (!IsConnected)
-            throw new InvalidOperationException($"The {nameof(BloomNode)} isn't connected to the remote server");
+        ThrowIfNotConnected();
 
         var player = new BloomPlayer(this, voiceChannel, textChannel);
         _players.Add(voiceChannel.GuildId, player);
-        await player.ConnectAsync(_config.SelfDeaf, _config.SelfMute);
+        await player.ConnectAsync(_configuration.SelfDeaf, _configuration.SelfMute);
         return player;
     }
 
-    public async Task LeaveAsync(IVoiceChannel voiceChannel)
+    /// <summary>
+    /// Leaves the player from the specified guild.
+    /// </summary>
+    /// <param name="guild">The guild where the player will be disconnected from.</param>
+    /// <returns>A <see cref="ValueTask"/> representing the asynchronous operation.</returns>
+    /// <exception cref="InvalidOperationException"></exception>
+    public async ValueTask LeaveAsync(IGuild guild)
     {
-        if (!IsConnected)
-            throw new InvalidOperationException($"The {nameof(BloomNode)} isn't connected to the remote server");
+        ThrowIfNotConnected();
 
-        BloomPlayer? player = GetPlayer(voiceChannel.Guild)
-            ?? throw new InvalidOperationException($"There is no player in {voiceChannel.Mention}");
+        if (!TryGetPlayer(guild, out BloomPlayer? player))
+            throw new InvalidOperationException("The player is not found in the specified guild");
 
-        _players.Remove(voiceChannel.GuildId);
-        await player.DisposeAsync();
-    }
-
-    public async Task LeaveAsync(IGuild guild)
-    {
-        if (!IsConnected)
-            throw new InvalidOperationException($"The {nameof(BloomNode)} isn't connected to the remote server");
-
-        BloomPlayer? player = GetPlayer(guild)
-            ?? throw new InvalidOperationException($"There is no player for {guild.Name}");
-
+        await DeletePlayerAsync(player);
+        await player.DisconnectAsync();
         _players.Remove(guild.Id);
-        await player.DisposeAsync();
     }
 
-    public async Task<SearchResult> SearchAsync(string query, SearchKind kind)
+    /// <summary>
+    /// Searches for tracks using the specified query and kind.
+    /// </summary>
+    /// <param name="query">The query to search for.</param>
+    /// <param name="kind">The kind of the search.</param>
+    /// <returns>A <see cref="LoadResult"/> as the result of the search.</returns>
+    /// <exception cref="InvalidOperationException"></exception>
+    public async ValueTask<LoadResult> SearchAsync(string query, SearchKind kind, CancellationToken cancellationToken = default)
     {
-        if (!IsConnected)
-            throw new InvalidOperationException($"The {nameof(BloomNode)} isn't connected to the remote server");
+        ThrowIfNotConnected();
 
-        query = kind switch
+        using HttpRequestMessage request = new(HttpMethod.Get, string.Format(LoadTracksEndpoint, kind.WrapQuery(query)));
+
+        using HttpResponseMessage response = await _restClient.SendAsync(request, cancellationToken);
+        if (!response.IsSuccessStatusCode)
         {
-            SearchKind.YouTube => $"ytsearch:{query}",
-            SearchKind.YouTubeMusic => $"ytmsearch:{query}",
-            SearchKind.SoundCloud => $"scsearch:{query}",
-            _ => query,
-        };
-
-        JsonNode data = await RequestAsync(HttpMethod.Get, string.Format(LoadTracksEndpoint, query));
-        return SearchResult.Parse(data);
-    }
-
-    internal async Task SendAsync(HttpMethod method, string uri, IPayload? payload = null)
-    {
-        using HttpRequestMessage requestMessage = new(method, $"{_config.RestEndpoint}{uri}");
-        if (payload is not null)
-        {
-            string jsonString = JsonSerializer.Serialize(payload, payload.GetType())
-                .Replace("\"null\"", "null");
-            requestMessage.Content = new StringContent(jsonString, Encoding.UTF8, "application/json");
+            string content = await response.Content.ReadAsStringAsync(cancellationToken);
+            NodeException?.Invoke(new NodeExceptionEventArgs($"Failed to search for tracks: {content}"));
         }
 
-        using HttpResponseMessage responseMessage = await _httpClient.SendAsync(requestMessage);
-        if (!responseMessage.IsSuccessStatusCode)
+        Stream stream = await response.Content.ReadAsStreamAsync(cancellationToken);
+        JsonNode? node = await JsonNode.ParseAsync(stream, cancellationToken: cancellationToken);
+        return ParseTool.ParseLoadResult(node);
+    }
+
+    internal async ValueTask UpdatePlayerAsync(BloomPlayer player, PlayerUpdatePayload updatePayload)
+    {
+        using HttpRequestMessage request = new(HttpMethod.Patch, string.Format(UpdatePlayerEndpoint, _sessionId, player.VoiceChannel.GuildId));
+        request.Content = JsonContent.Create(updatePayload);
+
+        using HttpResponseMessage response = await _restClient.SendAsync(request, _cancellationTokenSource!.Token);
+        if (!response.IsSuccessStatusCode)
         {
-            string response = await responseMessage.Content.ReadAsStringAsync();
-            NodeException?.Invoke(new NodeExceptionEventArgs(response));
+            string content = await response.Content.ReadAsStringAsync();
+            NodeException?.Invoke(new NodeExceptionEventArgs($"Failed to update player: {content}"));
         }
     }
 
-    internal async Task<JsonNode> RequestAsync(HttpMethod method, string uri, IPayload? payload = null)
+    internal async ValueTask DeletePlayerAsync(BloomPlayer player)
     {
-        using HttpRequestMessage requestMessage = new(method, $"{_config.RestEndpoint}{uri}");
-        if (payload is not null)
+        using HttpRequestMessage request = new(HttpMethod.Delete, string.Format(UpdatePlayerEndpoint, _sessionId, player.VoiceChannel.GuildId));
+
+        using HttpResponseMessage response = await _restClient.SendAsync(request, _cancellationTokenSource!.Token);
+        if (!response.IsSuccessStatusCode)
         {
-            string jsonString = JsonSerializer.Serialize(payload, payload.GetType())
-                .Replace("\"null\"", "null");
-            requestMessage.Content = new StringContent(jsonString, Encoding.UTF8, "application/json");
+            string content = await response.Content.ReadAsStringAsync();
+            NodeException?.Invoke(new NodeExceptionEventArgs($"Failed to delete player: {content}"));
         }
-
-        using HttpResponseMessage responseMessage = await _httpClient.SendAsync(requestMessage);
-        string response = await responseMessage.Content.ReadAsStringAsync();
-        if (!responseMessage.IsSuccessStatusCode)
-            NodeException?.Invoke(new NodeExceptionEventArgs(response));
-
-        return JsonNode.Parse(response)!;
-    }
-
-    private async Task ReconnectAsync()
-    {
-        if (0 <= _config.ReconnectAttemps || _config.ReconnectAttemps <= _reconnectAttemps)
-            return;
-
-        _connectionTokenSource.Cancel(false);
-        _reconnectAttemps++;
-        _reconnectDelay += _config.ReconnectDelayInMiliseconds;
-
-        Console.WriteLine($"Trying to connect in {_reconnectDelay} seconds");
-        await Task.Delay(_reconnectDelay);
-        await ConnectAsync();
     }
 
     private async Task StartReceivingAsync()
     {
-        while (_webSocket.State == WebSocketState.Open)
+        byte[] buffer = ArrayPool<byte>.Shared.Rent(8192);
+        while (IsConnected && !_cancellationTokenSource!.IsCancellationRequested)
         {
             try
             {
-                string jsonString = await ReceiveAsync();
-                JsonNode data = JsonNode.Parse(jsonString)!;
-                IGuild guild;
-                BloomPlayer player;
-                switch (data["op"]!.ToString())
-                {
-                    case "ready":
-                        SessionId = data["sessionId"]!.ToString();
-                        break;
-
-                    case "playerUpdate":
-                        if (bool.Parse(data["state"]!["connected"]!.ToString()))
-                        {
-                            guild = _discordClient.GetGuild(ulong.Parse(data["guildId"]!.ToString()));
-                            player = GetPlayer(guild)!;
-                            player.Track?.UpdatePosition(long.Parse(data["state"]!["position"]!.ToString()));
-                        }
-                        break;
-
-                    case "stats":
-                        break;
-
-                    case "event":
-                        guild = _discordClient.GetGuild(ulong.Parse(data["guildId"]!.ToString()));
-                        player = GetPlayer(guild)!;
-                        switch (data["type"]!.ToString())
-                        {
-                            case "TrackStartEvent":
-                                TrackStarted?.Invoke(new TrackStartEventArgs(player));
-                                break;
-
-                            case "TrackEndEvent":
-                                TrackEnded?.Invoke(new TrackEndEventArgs(player,
-                                    data["reason"]!.ToString() switch
-                                    {
-                                        "FINISHED" => TrackEndReason.Finished,
-                                        "LOAD_FAILED" => TrackEndReason.LoadFailed,
-                                        "STOPPED" => TrackEndReason.Stopped,
-                                        "REPLACED" => TrackEndReason.Replaced,
-                                        "CLEANUP" => TrackEndReason.Cleanup,
-                                        _ => TrackEndReason.None,
-                                    }
-                                ));
-                                break;
-
-                            case "TrackExceptionEvent":
-                                TrackException?.Invoke(new TrackExceptionEventArgs(player, new TrackEventException(
-                                    data["exception"]!["message"]!.ToString(),
-                                    data["exception"]!["cause"]!.ToString(),
-                                    Enum.Parse<TrackEventExceptionSeverity>(data["exception"]!["severity"]!.ToString(), ignoreCase: true)
-                                )));
-                                break;
-
-                            case "TrackStuckEvent":
-                                TrackStucked?.Invoke(new TrackStuckEventArgs(player, TimeSpan.FromMilliseconds(int.Parse(data["thresholdMs"]!.ToString()))));
-                                break;
-
-                            case "WebSocketClosedEvent":
-                                break;
-
-                            default:
-                                throw new Exception("Unhandled received event type");
-                        }
-                        break;
-
-                    default:
-                        throw new Exception("Unhandled received data type");
-                }
+                int length = await ReceiveDataAsync(buffer);
+                using var stream = new MemoryStream(buffer, 0, length);
+                JsonNode? node = await JsonNode.ParseAsync(stream, cancellationToken: _cancellationTokenSource.Token);
+                HandleReceivedData(node);
             }
-            catch
+            catch (OperationCanceledException)
             {
-                throw;
+            }
+            catch (Exception ex)
+            {
+                NodeException?.Invoke(new NodeExceptionEventArgs(ex.Message));
             }
         }
     }
 
-    private async ValueTask<string> ReceiveAsync()
+    private async ValueTask<int> ReceiveDataAsync(byte[] buffer)
     {
-        byte[] buffer = new byte[128 * 1024];
         int length = 0;
 
-        while (true)
+        while (IsConnected && !_cancellationTokenSource!.IsCancellationRequested)
         {
-            WebSocketReceiveResult result = await _webSocket.ReceiveAsync(new ArraySegment<byte>(buffer, length, 8192), CancellationToken.None);
+            WebSocketReceiveResult result = await _webSocketClient.ReceiveAsync(new ArraySegment<byte>(buffer, length, buffer.Length), _cancellationTokenSource.Token);
             if (result.MessageType == WebSocketMessageType.Close)
-            {
-                IsConnected = false;
                 break;
-            }
 
             length += result.Count;
             if (result.EndOfMessage)
                 break;
         }
 
-        return Encoding.UTF8.GetString(buffer, 0, length);
+        return length;
+    }
+
+    private void HandleReceivedData(JsonNode? node)
+    {
+        if (node is null)
+            return;
+
+        string op = node["op"]!.GetValue<string>();
+        if (op == "ready")
+        {
+            _sessionId = node["sessionId"]!.GetValue<string>();
+        }
+        else if (op == "playerUpdate")
+        {
+            bool connected = node["state"]!["connected"]!.GetValue<bool>();
+            if (connected)
+            {
+                ulong guildId = ulong.Parse(node["guildId"]!.GetValue<string>());
+                IGuild guild = _discordClient.GetGuild(guildId);
+                if (!TryGetPlayer(guild, out BloomPlayer? player))
+                    return;
+
+                if (player.Track is not null)
+                {
+                    long position = node["state"]!["position"]!.GetValue<long>();
+                    player.Track.Position = TimeSpan.FromMilliseconds(position);
+                }
+            }
+        }
+        else if (op == "event")
+        {
+            ulong guildId = ulong.Parse(node["guildId"]!.GetValue<string>());
+            IGuild guild = _discordClient.GetGuild(guildId);
+            if (!TryGetPlayer(guild, out BloomPlayer? player))
+                return;
+
+            string eventType = node["type"]!.GetValue<string>();
+            if (eventType == nameof(TrackStartEvent))
+            {
+                TrackStarted?.Invoke(new TrackStartEventArgs(player));
+            }
+            else if (eventType == nameof(TrackEndEvent))
+            {
+                TrackEnded?.Invoke(new TrackEndEventArgs(player, Enum.Parse<TrackEndReason>(node["reason"]!.GetValue<string>(), ignoreCase: true)));
+            }
+            else if (eventType == nameof(TrackExceptionEvent))
+            {
+                TrackException?.Invoke(new TrackExceptionEventArgs(player, ParseTool.ParseException(node["exception"]!)));
+            }
+            else if (eventType == nameof(TrackStuckEvent))
+            {
+                TrackStucked?.Invoke(new TrackStuckEventArgs(player, TimeSpan.FromMilliseconds(node["thresholdMs"]!.GetValue<long>())));
+            }
+        }
     }
 
     private async Task VoiceServerUpdated(SocketVoiceServer voiceServer)
     {
-        if (GetPlayer(voiceServer.Guild.Value) is BloomPlayer player)
+        if (!TryGetPlayer(voiceServer.Guild.Value, out BloomPlayer? player))
+            return;
+
+        await UpdatePlayerAsync(player, new PlayerUpdatePayload
         {
-            await player.UpdateSessionToConnectAsync(voiceServer);
-        }
+            Voice = new VoiceStatePayload
+            {
+                Token = voiceServer.Token,
+                Endpoint = voiceServer.Endpoint,
+                SessionId = player.VoiceSessionId,
+            }
+        });
     }
 
     private async Task UserVoiceStateUpdated(SocketUser user, SocketVoiceState before, SocketVoiceState after)
     {
-        if (user.Id == _discordClient.CurrentUser.Id)
-        {
-            if (after.VoiceChannel is not null)
-            {
-                BloomPlayer player = GetPlayer(after.VoiceChannel.Guild)!;
-                player.VoiceSessionId = after.VoiceSessionId;
-                if (player.VoiceChannel != after.VoiceChannel)
-                {
-                    await player.UpdateSessionIdAsync();
-                }
-            }
-            else if (HasPlayer(before.VoiceChannel.Guild))
-            {
-                await LeaveAsync(before.VoiceChannel);
-            }
-        }
-        else if (!user.IsBot)
-        {
-            if (before.VoiceChannel is null && after.VoiceChannel is not null)
-            {
-                _cancellationTokenSource?.Cancel();
-                _cancellationTokenSource?.Dispose();
-            }
-            else if (before.VoiceChannel is not null && after.VoiceChannel is null)
-            {
-                if (!HasPlayer(before.VoiceChannel.Guild))
-                    return;
+        SocketGuildUser guildUser = (user as SocketGuildUser)!;
+        if (!TryGetPlayer(guildUser.Guild, out BloomPlayer? player))
+            return;
 
-                if (before.VoiceChannel.Guild.VoiceChannels.Any(
-                        static (voiceChannel) => voiceChannel.ConnectedUsers.Any(
-                            static (connectedUser) => !connectedUser.IsBot
-                        )
-                    )
-                )
-                    return;
-
-                _cancellationTokenSource = new CancellationTokenSource();
-                _ = Task.Run(async () =>
-                {
-                    try
-                    {
-                        await Task.Delay(_config.LeaveDelayInMiliseconds, _cancellationTokenSource.Token);
-                        await LeaveAsync(before.VoiceChannel.Guild);
-                    }
-                    catch (OperationCanceledException)
-                    {
-                    }
-                }, _cancellationTokenSource.Token);
+        if (guildUser.Id == _discordClient.CurrentUser.Id)
+        {
+            if (after.VoiceChannel is null)
+            {
+                await LeaveAsync(before.VoiceChannel.Guild);
+                return;
             }
+
+            player.VoiceSessionId = after.VoiceSessionId;
+            player.VoiceChannel = after.VoiceChannel;
         }
+        else if (!guildUser.IsBot && before.VoiceChannel is not null && after.VoiceChannel is null)
+        {
+            if (guildUser.Guild.VoiceChannels.All(static (voiceChannel) => !voiceChannel.ConnectedUsers.Any(static (user) => !user.IsBot)))
+                await LeaveAsync(before.VoiceChannel.Guild);
+        }
+    }
+
+    private void ThrowIfConnected()
+    {
+        if (IsConnected)
+            throw new InvalidOperationException("The node already connected to the remote server");
+    }
+
+    private void ThrowIfNotConnected()
+    {
+        if (!IsConnected)
+            throw new InvalidOperationException("The node is not connected to the remote server");
     }
 }
