@@ -1,16 +1,25 @@
+using System;
+using System.Threading.Tasks;
+using Bloom.Playback;
+using Blossom.Services;
+using Discord;
+using Discord.Interactions;
+using Microsoft.Extensions.DependencyInjection;
+
 namespace Blossom.Preconditions;
 
 public sealed class RequireUserInVoiceChannelAttribute : PreconditionAttribute
 {
-    private const string UserNotInVoiceChannel = "You must be joined to {0}";
-
     public override Task<PreconditionResult> CheckRequirementsAsync(IInteractionContext context, ICommandInfo commandInfo, IServiceProvider services)
     {
-        BloomPlayer? player = services.GetRequiredService<BloomNode>().GetPlayer(context.Guild);
-        return Task.FromResult(
-            (player is not null && player.VoiceChannel != ((IVoiceState)context.User).VoiceChannel)
-                ? PreconditionResult.FromError(string.Format(UserNotInVoiceChannel, player.VoiceChannel.Mention))
-                : PreconditionResult.FromSuccess()
-        );
+        AudioService audioService = services.GetRequiredService<AudioService>();
+        BloomPlayer? player = audioService.GetPlayer(context.Guild);
+
+        if (player is not null && player.VoiceChannel != ((IVoiceState)context.User).VoiceChannel)
+        {
+            return Task.FromResult(PreconditionResult.FromError($"You must be joined to {player.VoiceChannel.Mention}"));
+        }
+
+        return Task.FromResult(PreconditionResult.FromSuccess());
     }
 }
