@@ -1,27 +1,11 @@
+using System.Threading.Tasks;
+using Bloom.Playback;
+
 namespace Bloom.Events;
 
+/// <summary>
+/// Represents the event that is called when an exception is thrown.
+/// </summary>
+/// <param name="args">The <see cref="TrackExceptionEventArgs"/> that contains the event data.</param>
+/// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
 public delegate Task TrackExceptionEvent(TrackExceptionEventArgs args);
-
-public sealed class TrackExceptionEventArgs : TrackEventArgs
-{
-    public TrackEventException Exception { get; }
-
-    internal TrackExceptionEventArgs(BloomPlayer player, TrackEventException exception) : base(player)
-    {
-        Exception = exception;
-    }
-}
-
-public sealed class TrackEventException
-{
-    public string Message { get; }
-    public string Cause { get; }
-    public TrackEventExceptionSeverity Severity { get; }
-
-    internal TrackEventException(string message, string cause, TrackEventExceptionSeverity severity)
-    {
-        Message = message;
-        Cause = cause;
-        Severity = severity;
-    }
-}
