@@ -2,13 +2,12 @@ namespace Bloom.Filters;
 
 /// <summary>
 /// Higher frequencies get suppressed, while lower frequencies pass through this filter, thus the name low pass.
-/// Any smoothing values equal to, or less than 1.0 will disable the filter.
+/// Any smoothing values less than or equal to 1.0 will disable the filter.
 /// </summary>
-public readonly struct LowPassFilter : IFilter
+public sealed class LowPassFilter : IFilter
 {
     /// <summary>
-    /// The smoothing factor (1.0 < )
+    /// The smoothing factor.
     /// </summary>
-    [JsonPropertyName("smoothing")]
     public float Smoothing { get; init; }
 }
