@@ -3,23 +3,20 @@ namespace Bloom.Filters;
 /// <summary>
 /// Changes the speed, pitch, and rate. All default to 1.0.
 /// </summary>
-public readonly struct TimescaleFilter : IFilter
+public sealed class TimescaleFilter : IFilter
 {
     /// <summary>
-    /// The playback speed (0.0 ≤).
+    /// The playback speed.
     /// </summary>
-    [JsonPropertyName("speed")]
     public float Speed { get; init; }
 
     /// <summary>
-    /// The pitch (0.0 ≤).
+    /// The pitch.
     /// </summary>
-    [JsonPropertyName("pitch")]
     public float Pitch { get; init; }
 
     /// <summary>
-    /// The rate (0.0 ≤).
+    /// The rate.
     /// </summary>
-    [JsonPropertyName("rate")]
     public float Rate { get; init; }
 }
