@@ -1,28 +1,52 @@
+using System;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+
 namespace Bloom.Payloads;
 
-internal sealed class PlayerUpdatePayload : IPayload
+internal sealed class PlayerUpdatePayload
 {
-    [JsonPropertyName("encodedTrack"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public string? EncodedTrack { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault), JsonConverter(typeof(EncodedTrackConverter))]
+    public string? EncodedTrack { get; init; }
 
-    [JsonPropertyName("identifier"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public string? Identifier { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public string? Identifier { get; init; }
 
-    [JsonPropertyName("position"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public long? Position { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public long? Position { get; init; }
 
-    [JsonPropertyName("endTime"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public long? EndTime { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public long? EndTime { get; init; }
 
-    [JsonPropertyName("volume"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public int? Volume { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public int? Volume { get; init; }
 
-    [JsonPropertyName("paused"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public bool? Paused { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public bool? Paused { get; init; }
 
-    [JsonPropertyName("filters"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public FilterPayload? Filters { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public FilterPayload? Filters { get; init; }
 
-    [JsonPropertyName("voice"), JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    public VoiceStatePayload? Voice { get; set; }
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
+    public VoiceStatePayload? Voice { get; init; }
+
+    private sealed class EncodedTrackConverter : JsonConverter<string>
+    {
+        public override string? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        {
+            return reader.TokenType switch
+            {
+                JsonTokenType.String => reader.GetString(),
+                JsonTokenType.Null => null,
+                _ => throw new JsonException(),
+            };
+        }
+
+        public override void Write(Utf8JsonWriter writer, string value, JsonSerializerOptions options)
+        {
+            // This is a workaround for stopping the playback of a track. Since we are ignoring null values,
+            // we need to write "null" as a string and this converter will convert it to a null value.
+            writer.WriteStringValue(value == "null" ? null : value);
+        }
+    }
 }
