@@ -1,3 +1,17 @@
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+using Bloom;
+using Bloom.Filters;
+using Bloom.Playback;
+using Bloom.Searching;
+using Blossom.AutoCompleteHandlers;
+using Blossom.Preconditions;
+using Blossom.Services;
+using Blossom.Utilities;
+using Discord;
+using Discord.Interactions;
+
 namespace Blossom.Modules;
 
 public sealed class AudioModule : BaseInteractionModule
@@ -9,11 +23,11 @@ public sealed class AudioModule : BaseInteractionModule
     private const string PositionBar = "─────────────────────────────";
     private const string PositionSign = "⚪";
 
-    private readonly BloomNode _bloomNode;
+    private readonly AudioService _audioService;
 
-    public AudioModule(IServiceProvider services, BloomNode bloomNode) : base(services)
+    public AudioModule(IServiceProvider services, AudioService audioService) : base(services)
     {
-        _bloomNode = bloomNode;
+        _audioService = audioService;
     }
 
     [SlashCommand("join", "Joins the voice channel")]
@@ -21,7 +35,7 @@ public sealed class AudioModule : BaseInteractionModule
     public async Task JoinCommand()
     {
         IVoiceChannel voiceChannel = VoiceState!.VoiceChannel;
-        await _bloomNode.JoinAsync(voiceChannel, Channel);
+        await _audioService.JoinAsync(voiceChannel, Channel);
         await RespondAsync($"Joined to {voiceChannel.Mention}");
     }
 
@@ -30,7 +44,7 @@ public sealed class AudioModule : BaseInteractionModule
     public async Task LeaveCommand()
     {
         IVoiceChannel voiceChannel = VoiceState!.VoiceChannel;
-        await _bloomNode.LeaveAsync(voiceChannel);
+        await _audioService.LeaveAsync(voiceChannel.Guild);
         await RespondAsync($"Leaved from {voiceChannel.Mention}");
     }
 
@@ -39,35 +53,37 @@ public sealed class AudioModule : BaseInteractionModule
     public async Task PlayCommand([Summary(description: "The track's URL or name to search")] string query)
     {
         string response = string.Empty;
-        BloomPlayer? player = _bloomNode.GetPlayer(Guild);
-        if (player is null)
-        {
-            player = await _bloomNode.JoinAsync(VoiceState!.VoiceChannel, Channel);
+        BloomPlayer player = await _audioService.GetOrJoinAsync(VoiceState!.VoiceChannel, Channel);
+        if (player.State is PlayerState.Connected)
             response = $"Joined to {player.VoiceChannel.Mention}\n";
-        }
 
-        query = query.Trim('<', '>');
-        bool isWellFormed = Uri.IsWellFormedUriString(query, UriKind.Absolute);
-        SearchResult result = await _bloomNode.SearchAsync(query, isWellFormed ? SearchKind.Direct : SearchKind.YouTube);
-        if (result.Kind is SearchResultKind.NoMatches or SearchResultKind.LoadFailed)
+        LoadResult result = await _audioService.SearchAsync(query);
+        if (result.Kind is LoadResultKind.Empty or LoadResultKind.Error)
         {
             await RespondAsync($"{response}Could't find any tracks with given query!");
             return;
         }
 
-        if (result.Kind is SearchResultKind.PlaylistLoaded)
+        if (result.Kind is LoadResultKind.Playlist)
         {
-            player.Queue.AddRange(result.Tracks!);
-            await RespondAsync($"{response}Playlist `{result.Playlist!.Name}` loaded with {result.Tracks!.Count} tracks.");
+            PlaylistLoadResult playlistResult = (PlaylistLoadResult)result;
+            player.Queue.Add(playlistResult.Tracks);
+            await RespondAsync($"{response}Playlist `{playlistResult.Name}` loaded with {playlistResult.Tracks.Count} tracks.");
+        }
+        else if (result.Kind is LoadResultKind.Track)
+        {
+            TrackLoadResult trackResult = (TrackLoadResult)result;
+            player.Queue.Add(trackResult.Track);
+            await RespondAsync($"{response}`{trackResult.Track.Title.EndAt(TrackTitleLimit)}` queued.");
         }
         else
         {
-            BloomTrack trackToAdd = result.Tracks![0];
-            player.Queue.Add(trackToAdd);
-            await RespondAsync($"{response}`{trackToAdd.Title.CutOff(TrackTitleLimit)}` queued.");
+            SearchLoadResult searchResult = (SearchLoadResult)result;
+            player.Queue.Add(searchResult.Tracks[0]);
+            await RespondAsync($"{response}`{searchResult.Tracks[0].Title.EndAt(TrackTitleLimit)}` queued.");
         }
 
-        if (player.State is PlayerState.None or PlayerState.Connected or PlayerState.Stopped)
+        if (player.State is PlayerState.Connected or PlayerState.Stopped)
             await player.PlayNextAsync();
     }
 
@@ -78,39 +94,45 @@ public sealed class AudioModule : BaseInteractionModule
         [Summary(description: "The position to be inserted"), MinValue(1)] int position
     )
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
 
-        query = query.Trim('<', '>');
-        bool isWellFormed = Uri.IsWellFormedUriString(query, UriKind.Absolute);
-        SearchResult result = await _bloomNode.SearchAsync(query, isWellFormed ? SearchKind.Direct : SearchKind.YouTube);
-        if (result.Kind is SearchResultKind.NoMatches or SearchResultKind.LoadFailed)
+        LoadResult result = await _audioService.SearchAsync(query);
+        if (result.Kind is LoadResultKind.Empty or LoadResultKind.Error)
         {
             await RespondAsync("Could't find any tracks with given query!");
             return;
         }
 
-        position = Math.Min(position - 1, player.Queue.Count);
-        if (result.Kind is SearchResultKind.PlaylistLoaded)
+        position = Math.Min(position, player.Queue.Count) - 1;
+
+        if (result.Kind is LoadResultKind.Playlist)
         {
-            foreach (BloomTrack trackToInsert in result.Tracks!.Reverse())
+            PlaylistLoadResult playlistResult = (PlaylistLoadResult)result;
+            foreach (BloomTrack trackToInsert in playlistResult.Tracks.Reverse())
             {
                 player.Queue.InsertAt(trackToInsert, position);
             }
-            await RespondAsync($"Playlist `{result.Playlist!.Name}` loaded with {result.Tracks!.Count} tracks.");
+            await RespondAsync($"Playlist `{playlistResult.Name}` loaded with {playlistResult.Tracks.Count} tracks.");
+        }
+        else if (result.Kind is LoadResultKind.Track)
+        {
+            TrackLoadResult trackResult = (TrackLoadResult)result;
+            player.Queue.InsertAt(trackResult.Track, position);
+            await RespondAsync($"`{trackResult.Track.Title.EndAt(TrackTitleLimit)}` queued.");
         }
         else
         {
-            BloomTrack trackToInsert = result.Tracks![0];
-            player.Queue.InsertAt(trackToInsert, position);
-            await RespondAsync($"`{trackToInsert.Author} - {trackToInsert.Title.CutOff(TrackTitleLimit)}` queued.");
+            SearchLoadResult searchResult = (SearchLoadResult)result;
+            player.Queue.InsertAt(searchResult.Tracks[0], position);
+            await RespondAsync($"`{searchResult.Tracks[0].Title.EndAt(TrackTitleLimit)}` queued.");
         }
     }
 
     [SlashCommand("remove", "Removes a track from the queue")]
     [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
-    public async Task RemoveCommand([Summary(description: "The position of track"), MinValue(1), AutoComplete<RemoveTrackAutocompleteHandler>()] int position)
+    public async Task RemoveCommand([Summary(description: "The position of track"), MinValue(1), Autocomplete(typeof(RemoveTrackAutocompleteHandler))] int position)
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         if (position-- > player.Queue.Count)
         {
             await RespondAsync($"The provided positon must be lesser than or equal to queue's track count ({player.Queue.Count})!");
@@ -124,33 +146,33 @@ public sealed class AudioModule : BaseInteractionModule
         }
 
         BloomTrack removedTrack = player.Queue.RemoveAt(position);
-        await RespondAsync($"`{removedTrack.Title.CutOff(TrackTitleLimit)}` removed from the queue.");
+        await RespondAsync($"`{removedTrack.Title.EndAt(TrackTitleLimit)}` removed from the queue.");
     }
 
     [SlashCommand("stop", "Stops the player and clears the queue")]
     [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
     public async Task StopCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         await player.StopAsync();
         player.Queue.Clear();
         await RespondAsync("Player stopped and queue cleared.");
     }
 
     [SlashCommand("pause", "Pauses the player")]
-    [RequirePlayerJoined, RequireUserInVoiceChannel, EnsurePlayerState(PlayerState.Playing)]
+    [RequirePlayerJoined, RequireUserInVoiceChannel, RequirePlayerState(PlayerState.Playing)]
     public async Task PauseCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         await player.PauseAsync();
         await RespondAsync("Player paused.");
     }
 
     [SlashCommand("resume", "Resumes the player")]
-    [RequirePlayerJoined, RequireUserInVoiceChannel, EnsurePlayerState(PlayerState.Paused)]
+    [RequirePlayerJoined, RequireUserInVoiceChannel, RequirePlayerState(PlayerState.Paused)]
     public async Task ResumeCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         await player.ResumeAsync();
         await RespondAsync("Player resumed.");
     }
@@ -159,16 +181,16 @@ public sealed class AudioModule : BaseInteractionModule
     [RequirePlayerJoined, RequireUserInVoiceChannel]
     public async Task VolumeCommand([Summary(description: "The volume value"), MinValue(0), MaxValue(200)] ushort volume)
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         await player.SetVolumeAsync(volume);
         await RespondAsync($"Volume setted to `{volume}`.");
     }
 
     [SlashCommand("seek", "Seeks the current track to given position (mm:ss)")]
-    [RequirePlayerJoined, RequireUserInVoiceChannel, EnsurePlayerState(PlayerState.Playing)]
+    [RequirePlayerJoined, RequireUserInVoiceChannel, RequirePlayerState(PlayerState.Playing)]
     public async Task SeekAsync([Summary(description: "The position to seek")] TimeSpan position)
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
 
         if (player.Track!.Duration.CompareTo(position) < 1)
         {
@@ -185,16 +207,16 @@ public sealed class AudioModule : BaseInteractionModule
     [RequirePlayerJoined, RequireUserInVoiceChannel]
     public async Task LoopCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
-        player.Queue.Loop();
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
+        player.Queue.LoopMode = player.Queue.LoopMode.Next();
         await RespondAsync($"Loop mode set to: `{player.Queue.LoopMode}`");
     }
 
     [SlashCommand("next", "Stops the current track and plays next one if exists")]
-    [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue, EnsureNotLoopMode(LoopMode.One)]
+    [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
     public async Task NextCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         string response = "There must be some bug!";
         if (player.State is PlayerState.Playing or PlayerState.Paused)
         {
@@ -212,10 +234,10 @@ public sealed class AudioModule : BaseInteractionModule
     }
 
     [SlashCommand("previous", "Stops the current track and plays previous one if exists")]
-    [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue, EnsureNotLoopMode(LoopMode.One)]
+    [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
     public async Task PreviousCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         string response = "There must be some bug!";
         if (player.State is PlayerState.Playing or PlayerState.Paused)
         {
@@ -236,7 +258,7 @@ public sealed class AudioModule : BaseInteractionModule
     [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
     public async Task RewindCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         if (player.State is PlayerState.Playing)
             await player.SeekAsync(TimeSpan.Zero);
         else
@@ -247,25 +269,25 @@ public sealed class AudioModule : BaseInteractionModule
 
     [SlashCommand("filter", "Applies a filter preset to playback")]
     [RequirePlayerJoined, RequireUserInVoiceChannel]
-    public async Task FilterCommand([Summary(description: "The name of the preset"), AutoComplete<FilterAutoCompleteHandler>()] string name)
+    public async Task FilterCommand([Summary(description: "The name of the preset"), Autocomplete(typeof(FilterAutocompleteHandler))] string name)
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
-        FilterPreset? preset = AudioService.GetFilterPreset(name);
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
+        FilterPreset? preset = _audioService.GetFilterPreset(name);
         if (preset is null)
         {
             await RespondAsync("Preset couldn't found with provided name!");
             return;
         }
 
-        await player.ApplyFiltersAsync(preset.Filters, preset.Volume, preset.Bands);
-        await RespondAsync($"{preset.Name} applied.");
+        await player.ApplyFiltersAsync(preset.Filters, preset.Volume, preset.Equalizer);
+        await RespondAsync($"{preset.Name} applied. It may take a few seconds to change.");
     }
 
     [SlashCommand("shuffle", "Shuffles the queue")]
     [RequirePlayerJoined, RequireUserInVoiceChannel, RequireNonEmptyQueue]
     public async Task ShuffleCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         player.Queue.Shuffle();
         await RespondAsync("Queue shuffled.");
     }
@@ -274,66 +296,78 @@ public sealed class AudioModule : BaseInteractionModule
     [RequirePlayerJoined, RequireNonEmptyQueue]
     public async Task QueueCommand([Summary(description: "The page number of the queue (zero-indexed)")] int? page = null)
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         page ??= player.Queue.Current / TrackPerPage;
         int startIndex = page.Value * TrackPerPage;
         BloomTrack[] tracks = player.Queue.Take(startIndex..(startIndex + TrackPerPage)).Where((track) => track is not null).ToArray();
-        await RespondWithEmbedAsync(
+
+        Embed embed = EmbedUtility.CreateEmbed(
             title: $"-- Queue --",
             description: string.Join('\n', tracks
                     .Select((track, index) =>
                         $"{startIndex + index + 1}) [{track.Title}]({track.Url}){((startIndex + index == player.Queue.Current) ? " (Now Playing)" : string.Empty)}"
                     )
                 ),
-            footer: CreateFooter($"Page {page}")
+            color: Cherry,
+            footer: EmbedUtility.CreateFooter($"Page {page}")
         );
+
+        await RespondAsync(embed: embed);
     }
 
     [SlashCommand("nowplaying", "Shows the currently playing track's information")]
-    [RequirePlayerJoined, EnsurePlayerState(PlayerState.Playing)]
+    [RequirePlayerJoined, RequirePlayerState(PlayerState.Playing)]
     public async Task NowPlayingCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
 
         string description = $"[{player.Track!.Title}]({player.Track.Url})\n";
         if (player.Track.IsStream)
         {
-            await RespondWithEmbedAsync(
-                title: NowPlayingTitle,
-                description: $"{description}```\n{StreamDescription}```"
-            );
-            return;
+            description += $"{StreamDescription}\n";
+        }
+        else
+        {
+            description += $"```\n{PositionBar.Insert((int)(player.Track.Position.TotalSeconds / player.Track.Duration.TotalSeconds * 30), PositionSign)}\n\n";
+            description += $"[ {player.Track.Position:hh\\:mm\\:ss} / {player.Track.Duration:hh\\:mm\\:ss} ]```";
         }
 
-        await RespondWithEmbedAsync(
+        Embed embed = EmbedUtility.CreateEmbed(
             title: NowPlayingTitle,
-            description: description +
-                $"```\n{PositionBar.Insert((int)(player.Track.Position.TotalSeconds / player.Track.Duration.TotalSeconds * 30), PositionSign)}\n\n"
-              + $"[ {player.Track.Position:hh\\:mm\\:ss} / {player.Track.Duration:hh\\:mm\\:ss} ]```"
+            description: description,
+            thumbnail: player.Track.ArtworkUrl,
+            color: Cherry
         );
+
+        await RespondAsync(embed: embed);
     }
 
     [SlashCommand("lyrics", "Shows the lyrics for current track")]
     [RequirePlayerJoined]
     public async Task LyricsCommand()
     {
-        BloomPlayer player = _bloomNode.GetPlayer(Guild)!;
+        await DeferAsync();
+
+        BloomPlayer player = _audioService.GetPlayer(Guild)!;
         if (player.Track is null || player.State is not PlayerState.Playing or PlayerState.Paused)
         {
-            await RespondAsync("Nothing is playing right now!");
+            await FollowupAsync("Nothing is playing right now!");
             return;
         }
 
-        string? lyrics = await player.Track.GetLyricsAsync();
+        LyricsResult? lyrics = await _audioService.GetLyricsAsync(player.Track);
         if (lyrics is null)
         {
-            await RespondAsync("Couldn't find lyrics for this song!");
+            await FollowupAsync("Couldn't find lyrics for this song!");
             return;
         }
 
-        await RespondWithEmbedAsync(
-            title: $"Lyrics for {player.Track.Title.CutOff(TrackTitleLimit)}",
-            description: lyrics
+        Embed embed = EmbedUtility.CreateEmbed(
+            title: $"Lyrics for {player.Track.Title.EndAt(TrackTitleLimit)}",
+            description: lyrics.Lyrics,
+            color: Cherry
         );
+
+        await FollowupAsync(embed: embed);
     }
 }
