@@ -1,15 +1,23 @@
+using System;
+using System.Threading.Tasks;
+using Blossom.Services;
+using Discord;
+using Discord.Interactions;
+using Microsoft.Extensions.DependencyInjection;
+
 namespace Blossom.Preconditions;
 
 public sealed class RequirePlayerJoinedAttribute : PreconditionAttribute
 {
-    private const string NotJoinedMessage = "I'm not joined to voice a channel!";
-
     public override Task<PreconditionResult> CheckRequirementsAsync(IInteractionContext context, ICommandInfo commandInfo, IServiceProvider services)
     {
-        return Task.FromResult(
-            !services.GetRequiredService<BloomNode>().HasPlayer(context.Guild)
-                ? PreconditionResult.FromError(NotJoinedMessage)
-                : PreconditionResult.FromSuccess()
-        );
+        AudioService audioService = services.GetRequiredService<AudioService>();
+
+        if (!audioService.HasPlayer(context.Guild))
+        {
+            return Task.FromResult(PreconditionResult.FromError("I'm not joined to voice a channel!"));
+        }
+
+        return Task.FromResult(PreconditionResult.FromSuccess());
     }
 }
