@@ -1,15 +1,21 @@
+using System;
+using System.Threading.Tasks;
+using Discord;
+using Discord.Interactions;
+
 namespace Blossom.Preconditions;
 
 public sealed class RequireVoiceChannelAttribute : PreconditionAttribute
 {
-    private const string NoVoiceChannel = "You must be joined to a voice channel!";
-
     public override Task<PreconditionResult> CheckRequirementsAsync(IInteractionContext context, ICommandInfo commandInfo, IServiceProvider services)
     {
-        return Task.FromResult(
-            (((IVoiceState)context.User).VoiceChannel is null)
-                ? PreconditionResult.FromError(NoVoiceChannel)
-                : PreconditionResult.FromSuccess()
-        );
+        IVoiceState voiceState = (IVoiceState)context.User;
+
+        if (voiceState.VoiceChannel is null)
+        {
+            return Task.FromResult(PreconditionResult.FromError("You must be joined to a voice channel!"));
+        }
+
+        return Task.FromResult(PreconditionResult.FromSuccess());
     }
 }
